requirements.txt
fastapi
pymongo
requests
passlib
uvicorn[standard]
python-jose
python-jose[cryptography]
passlib[bcrypt] 
pyjwt
python-multipart
python-dotenv
pydantic[email]